move MSA3D to src/msa3d, few other changes
