--- environment.yml
@@ -1,215 +1,0 @@
-name: jwst_update_1-14
-channels:
-  - defaults
-dependencies:
-  - anyio=4.2.0=py311hecd8cb5_0
-  - appnope=0.1.2=py311hecd8cb5_1001
-  - argon2-cffi=21.3.0=pyhd3eb1b0_0
-  - argon2-cffi-bindings=21.2.0=py311h6c40b1e_0
-  - asttokens=2.0.5=pyhd3eb1b0_0
-  - async-lru=2.0.4=py311hecd8cb5_0
-  - babel=2.11.0=py311hecd8cb5_0
-  - beautifulsoup4=4.12.2=py311hecd8cb5_0
-  - bleach=4.1.0=pyhd3eb1b0_0
-  - bzip2=1.0.8=h1de35cc_0
-  - ca-certificates=2023.12.12=hecd8cb5_0
-  - certifi=2024.2.2=py311hecd8cb5_0
-  - cffi=1.16.0=py311h6c40b1e_0
-  - comm=0.1.2=py311hecd8cb5_0
-  - debugpy=1.6.7=py311hcec6c5f_0
-  - decorator=5.1.1=pyhd3eb1b0_0
-  - defusedxml=0.7.1=pyhd3eb1b0_0
-  - executing=0.8.3=pyhd3eb1b0_0
-  - ipykernel=6.28.0=py311hecd8cb5_0
-  - ipython=8.20.0=py311hecd8cb5_0
-  - jedi=0.18.1=py311hecd8cb5_1
-  - jinja2=3.1.3=py311hecd8cb5_0
-  - json5=0.9.6=pyhd3eb1b0_0
-  - jupyter-lsp=2.2.0=py311hecd8cb5_0
-  - jupyter_client=8.6.0=py311hecd8cb5_0
-  - jupyter_core=5.5.0=py311hecd8cb5_0
-  - jupyter_events=0.8.0=py311hecd8cb5_0
-  - jupyter_server=2.10.0=py311hecd8cb5_0
-  - jupyter_server_terminals=0.4.4=py311hecd8cb5_1
-  - jupyterlab=4.0.11=py311hecd8cb5_0
-  - jupyterlab_pygments=0.1.2=py_0
-  - jupyterlab_server=2.25.1=py311hecd8cb5_0
-  - libcxx=14.0.6=h9765a3e_0
-  - libffi=3.4.4=hecd8cb5_0
-  - libsodium=1.0.18=h1de35cc_0
-  - markupsafe=2.1.3=py311h6c40b1e_0
-  - matplotlib-inline=0.1.6=py311hecd8cb5_0
-  - mistune=2.0.4=py311hecd8cb5_0
-  - nbclient=0.8.0=py311hecd8cb5_0
-  - nbconvert=7.10.0=py311hecd8cb5_0
-  - nbformat=5.9.2=py311hecd8cb5_0
-  - ncurses=6.4=hcec6c5f_0
-  - nest-asyncio=1.6.0=py311hecd8cb5_0
-  - notebook=7.0.8=py311hecd8cb5_0
-  - notebook-shim=0.2.3=py311hecd8cb5_0
-  - openssl=3.0.13=hca72f7f_0
-  - overrides=7.4.0=py311hecd8cb5_0
-  - pandocfilters=1.5.0=pyhd3eb1b0_0
-  - parso=0.8.3=pyhd3eb1b0_0
-  - pexpect=4.8.0=pyhd3eb1b0_3
-  - pip=23.3.1=py311hecd8cb5_0
-  - platformdirs=3.10.0=py311hecd8cb5_0
-  - prometheus_client=0.14.1=py311hecd8cb5_0
-  - prompt-toolkit=3.0.43=py311hecd8cb5_0
-  - prompt_toolkit=3.0.43=hd3eb1b0_0
-  - ptyprocess=0.7.0=pyhd3eb1b0_2
-  - pure_eval=0.2.2=pyhd3eb1b0_0
-  - pycparser=2.21=pyhd3eb1b0_0
-  - pygments=2.15.1=py311hecd8cb5_1
-  - python=3.11.7=hf27a42d_0
-  - python-dateutil=2.8.2=pyhd3eb1b0_0
-  - python-fastjsonschema=2.16.2=py311hecd8cb5_0
-  - python-json-logger=2.0.7=py311hecd8cb5_0
-  - pytz=2023.3.post1=py311hecd8cb5_0
-  - pyyaml=6.0.1=py311h6c40b1e_0
-  - pyzmq=25.1.2=py311hcec6c5f_0
-  - readline=8.2=hca72f7f_0
-  - requests=2.31.0=py311hecd8cb5_1
-  - rfc3339-validator=0.1.4=py311hecd8cb5_0
-  - rfc3986-validator=0.1.1=py311hecd8cb5_0
-  - send2trash=1.8.2=py311hecd8cb5_0
-  - setuptools=68.2.2=py311hecd8cb5_0
-  - six=1.16.0=pyhd3eb1b0_1
-  - sniffio=1.3.0=py311hecd8cb5_0
-  - soupsieve=2.5=py311hecd8cb5_0
-  - sqlite=3.41.2=h6c40b1e_0
-  - stack_data=0.2.0=pyhd3eb1b0_0
-  - terminado=0.17.1=py311hecd8cb5_0
-  - tinycss2=1.2.1=py311hecd8cb5_0
-  - tk=8.6.12=h5d9f67b_0
-  - tornado=6.3.3=py311h6c40b1e_0
-  - traitlets=5.7.1=py311hecd8cb5_0
-  - typing-extensions=4.9.0=py311hecd8cb5_1
-  - typing_extensions=4.9.0=py311hecd8cb5_1
-  - wcwidth=0.2.5=pyhd3eb1b0_0
-  - webencodings=0.5.1=py311hecd8cb5_1
-  - websocket-client=0.58.0=py311hecd8cb5_4
-  - wheel=0.41.2=py311hecd8cb5_0
-  - xz=5.4.5=h6c40b1e_0
-  - yaml=0.2.5=haf1e3a3_0
-  - zeromq=4.3.5=hcec6c5f_0
-  - zlib=1.2.13=h4dc903c_0
-  - pip:
-      - alabaster==0.7.16
-      - asdf==3.2.0
-      - asdf-astropy==0.5.0
-      - asdf-coordinates-schemas==0.3.0
-      - asdf-standard==1.1.1
-      - asdf-transform-schemas==0.5.0
-      - asdf-unit-schemas==0.1.0
-      - asdf-wcs-schemas==0.4.0
-      - astro-prospector==1.3.0
-      - astro-sedpy==0.3.2
-      - astropy==6.0.0
-      - astropy-iers-data==0.2024.2.26.0.28.55
-      - astroquery==0.4.7
-      - astroscrappy==1.2.0
-      - attrs==23.2.0
-      - backports-tarfile==1.2.0
-      - bayesicfitting==3.2.0
-      - blinker==1.8.2
-      - charset-normalizer==3.3.2
-      - click==8.1.7
-      - contourpy==1.2.0
-      - crds==11.17.19
-      - cycler==0.12.1
-      - cython==3.0.10
-      - docutils==0.21.2
-      - drizzle==1.15.1
-      - dust-extinction==1.4.1
-      - eazy==0.6.8
-      - et-xmlfile==1.1.0
-      - extinction==0.4.6
-      - filelock==3.13.1
-      - flask==3.0.3
-      - fonttools==4.49.0
-      - future==1.0.0
-      - geos==0.2.3
-      - grizli==1.11.9
-      - gwcs==0.21.0
-      - h5py==3.11.0
-      - html5lib==1.1
-      - idna==3.6
-      - imageio==2.34.1
-      - imagesize==1.4.1
-      - importlib-metadata==7.0.1
-      - itsdangerous==2.2.0
-      - jaraco-classes==3.4.0
-      - jaraco-context==5.3.0
-      - jaraco-functools==4.0.1
-      - jmespath==1.0.1
-      - joblib==1.4.2
-      - jsonschema==4.21.1
-      - jsonschema-specifications==2023.12.1
-      - jwst==1.14.0
-      - keyring==25.2.1
-      - kiwisolver==1.4.5
-      - lazy-loader==0.4
-      - lxml==5.2.2
-      - mastquery==1.8.2
-      - matplotlib==3.8.3
-      - more-itertools==10.3.0
-      - msaexp==0.8.5
-      - networkx==3.3
-      - nsclean==1.5
-      - numpy==1.25.2
-      - numpydoc==1.7.0
-      - opencv-python-headless==4.9.0.80
-      - openpyxl==3.1.4
-      - packaging==23.2
-      - pandas==2.2.2
-      - parsley==1.3
-      - peakutils==1.3.4
-      - photutils==1.11.0
-      - pillow==10.2.0
-      - poppy==1.1.1
-      - psutil==5.9.8
-      - pyerfa==2.0.1.1
-      - pyparsing==3.1.1
-      - pysiaf==0.22.0
-      - pyvo==1.5.2
-      - referencing==0.33.0
-      - regions==0.9
-      - rpds-py==0.18.0
-      - scikit-image==0.23.2
-      - scikit-learn==1.5.0
-      - scipy==1.9.3
-      - semantic-version==2.10.0
-      - sep==1.2.1
-      - shapely==2.0.4
-      - snowballstemmer==2.2.0
-      - spherical-geometry==1.3.1
-      - sphinx==7.3.7
-      - sphinxcontrib-applehelp==1.0.8
-      - sphinxcontrib-devhelp==1.0.6
-      - sphinxcontrib-htmlhelp==2.0.5
-      - sphinxcontrib-jsmath==1.0.1
-      - sphinxcontrib-qthelp==1.0.7
-      - sphinxcontrib-serializinghtml==1.1.10
-      - sregion==1.3.2
-      - stcal==1.7.1
-      - stdatamodels==1.10.1
-      - stpipe==0.5.2
-      - stsci-image==2.3.5
-      - stsci-imagestats==1.6.3
-      - stsci-stimage==0.2.6
-      - stsci-tools==4.1.0
-      - stwcs==1.7.3
-      - synphot==1.4.0
-      - tabulate==0.9.0
-      - threadpoolctl==3.5.0
-      - tifffile==2024.5.22
-      - tqdm==4.66.4
-      - tristars==0.1
-      - tweakwcs==0.8.6
-      - tzdata==2024.1
-      - urllib3==2.2.1
-      - werkzeug==3.0.3
-      - wiimatch==0.3.2
-      - zipp==3.17.0
-prefix: /Users/ibarisic/opt/anaconda3/envs/jwst_update_1-14
