+++ pyproject.toml
@@ -1,11 +1,55 @@
+[project]
+name = "msa3d"
+description = "data reduction and cube design for the JWST slit-stepping survey GO-2136 and future surveys"
+license = {file = "LICENSE"}
+readme = "README.md"
+#version = "0.0.1"
+authors = [
+    {name = "Ivana Barisic", email = "ibarisic@ucdavis.edu"}
+]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "License :: OSI Approved :: BSD-3 License",
+    "Intended Audience :: Developers",
+    "Intended Audience :: Science/Research",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python",
+]
+dynamic = ["version"]
+requires-python = ">=3.10"
+dependencies = [
+    "astropy",
+    "astroscrappy",
+    "h5py",
+    "jupyter",
+    "jupyterlab",
+    "jwst",
+    "msaexp",
+    "numpy<2.0",
+    "pandas",
+    "scipy",
+    
+]
+[project.urls]
+"Source Code" = "https://github.com/barisiciv/msa3d"
+
+
 [build-system]
 
 requires = [
     "setuptools",
     "setuptools_scm",
     "wheel",
-    "oldest-supported-numpy",
-    "cython>=0.29.14"
+    #"oldest-supported-numpy",
+    #"cython>=0.29.14"
 ]
 
-build-backend = 'setuptools.build_meta'
+[tools.setuptools_scm]
+write_to = "src/msa3d/version.py"
+
+#stuff to include package data, if needed
+[tool.setuptools.package-data]
+"*" = ["*.fits"]
+"msa3d.processing" = ["*.fits"]
+
+#build-backend = 'setuptools.build_meta'
