--- setup.cfg
@@ -1,63 +1,0 @@
-[metadata]
-name = msa3d
-description = Processing and datacube construction routines for JWST MSA slit-stepping observations
-long_description = file: README.rst
-long_description_content_type = text/x-rst
-author = Ivana Barisic
-author_email = ibarisic@ucdavis.edu
-license = BSD-3
-url = https://github.com/barisiciv/msa3d
-edit_on_github = False
-github_project = barisiciv/msa3d
-keywords = JWST, MSA, slit-stepping, datacube, data reduction, spectroscopy
-classifiers =
-    Development Status :: 4 - Beta
-    Intended Audience :: Science/Research
-    License :: OSI Approved :: BSD License
-    Natural Language :: English
-    Operating System :: OS Independent
-    Programming Language :: Python
-    Programming Language :: Python :: 3.11
-    Programming Language :: Python :: 3.12
-    Topic :: Documentation :: Sphinx
-    Topic :: Scientific/Engineering :: Astronomy
-    Topic :: Software Development :: Libraries :: Python Modules
-    Topic :: Software Development :: User Interfaces
-
-[options]
-zip_safe = False
-use_2to3=False
-packages = find:
-python_requires = >=3.11,<3.13
-setup_requires = setuptools_scm
-include_package_data = True
-install_requires =
-    numpy>=1.24,<2.0
-    matplotlib>=3.8
-    scipy>=1.9
-    pandas>=2.2
-    pillow>=10.2
-    astropy>=6.0
-    jwst==1.14.0
-    astroscrappy==1.2.0
-    grizli>=1.11
-    msaexp==0.8.5
-#scripts =
-#    bin/run_msa3d
-
-[options.extras_require]
-cupy =
-    cupy>=13.0
-#test =
-#    pytest>=7.0.0
-#    pytest-astropy
-#    tox
-#    pytest-cov
-#    coverage
-#docs =
-#    sphinx>=1.6,<8
-#    docutils<0.21
-#    sphinx-automodapi
-#    sphinx_rtd_theme==2.0.0
-
-
