+++ src/msa3d/pre_processing/stsci_stage1.py
