+++ src/msa3d/processing/jwst_nirspec_pathloss_0002.fits
