+++ src/msa3d/processing/stsci_stage2.py
